Refactored to use floating point coordinates. This should be postscript points with 72 to an inch.

+++ src/org/pathwayeditor/businessobjects/typedefn/ILabelAttributeDefaults.java
@@ -28,5 +28,5 @@ public interface ILabelAttributeDefaults {
 
 	RGB getFillColour();
 
-	int getLineWidth();
+	double getLineWidth();
 }
